+++ Group11-main/Group11-main/Routes/User.Routes.js
@@ -1,0 +1,7 @@
+//@degas123
+const Users = require("../Controllers/User.Controler");
+const auth = require("../../Lib/Middleware");
+
+module.exports= function(app){
+    app.route("/users").post(Users.create_account);
+}
